Extract resumable projection common code to trait
