+++ common/src/main/scala/hmda/messages/CommonMessages.scala
@@ -3,4 +3,7 @@ package hmda.messages
 object CommonMessages {
   sealed trait Message
   final case object StopActor extends Message
+
+  trait Command
+  trait Event
 }
