+++ common/src/main/scala/hmda/messages/institution/InstitutionEvents.scala
@@ -1,10 +1,11 @@
 package hmda.messages.institution
 
+import hmda.messages.CommonMessages.Event
 import hmda.model.institution.Institution
 
 object InstitutionEvents {
 
-  sealed trait InstitutionEvent
+  sealed trait InstitutionEvent extends Event
   case class InstitutionCreated(i: Institution) extends InstitutionEvent
   case class InstitutionModified(i: Institution) extends InstitutionEvent
   case class InstitutionDeleted(LEI: String) extends InstitutionEvent
