+++ institutions-api/src/main/scala/hmda/institution/HmdaInstitutionApi.scala
@@ -5,7 +5,7 @@ import com.typesafe.config.ConfigFactory
 import hmda.institution.api.http.HmdaInstitutionQueryApi
 import org.slf4j.LoggerFactory
 import akka.actor.typed.scaladsl.adapter._
-import hmda.institution.projection.InstitutionDBProjector
+import hmda.institution.projection.InstitutionDBProjection
 import hmda.messages.projection.CommonProjectionMessages.StartStreaming
 
 object HmdaInstitutionApi extends App {
@@ -31,7 +31,7 @@ object HmdaInstitutionApi extends App {
 
   system.actorOf(HmdaInstitutionQueryApi.props(), "hmda-institutions-api")
   val institutionDBProjector =
-    system.spawn(InstitutionDBProjector.behavior, InstitutionDBProjector.name)
+    system.spawn(InstitutionDBProjection.behavior, InstitutionDBProjection.name)
   institutionDBProjector ! StartStreaming
 
 }
